examples/tavus-livekit/pyproject.toml
[project]
name = "tavus-livekit"
version = "0.1.0"
description = "Tavus Avatar Agent powered by LiveKit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[tavus,openai,deepgram,cartesia]~=1.3",
    "python-dotenv>=1.0.0",
]

